refactor(frontend): extract common library
