+++ .moon/workspace.yml
@@ -7,6 +7,7 @@ projects:
   importer: 'apps/importer'
   generated: 'libs/generated'
   graphql: 'libs/graphql'
+  utilities: 'libs/utilities'
 
 vcs:
   manager: 'git'
