+++ apps/frontend/src/lib/utilities.ts
@@ -12,26 +12,7 @@ import {
 	IconHeadphones,
 	IconMicrophone,
 } from "@tabler/icons-react";
-import { camelCase, startCase } from "lodash";
 import { match } from "ts-pattern";
-
-/**
- * Generate initials for a given string.
- */
-export const getInitials = (name: string) => {
-	const rgx = new RegExp(/(\p{L}{1})\p{L}+/, "gu");
-	const initials = [...name.matchAll(rgx)] || [];
-	const actuals = (
-		(initials.shift()?.[1] || "") + (initials.pop()?.[1] || "")
-	).toUpperCase();
-	return actuals;
-};
-
-/**
- * Change case to a presentable format
- */
-export const changeCase = (name: string) =>
-	startCase(camelCase(name.toLowerCase()));
 
 /**
  * Get the correct name of the lot from a string
