+++ libs/utilities/moon.yml
@@ -1,0 +1,8 @@
+type: library
+
+workspace:
+  inheritedTasks:
+    exclude:
+      - build
+      - lint
+      - format
