+++ libs/utilities/package.json
@@ -1,0 +1,9 @@
+{
+	"name": "@ryot/utilities",
+	"dependencies": {
+		"lodash": "4.17.21"
+	},
+	"devDependencies": {
+		"@types/lodash": "4.14.194"
+	}
+}
