+++ libs/utilities/tsconfig.json
@@ -1,0 +1,13 @@
+{
+	"extends": "../../tsconfig.options.json",
+	"compilerOptions": {
+		"module": "esnext",
+		"moduleResolution": "node",
+		"verbatimModuleSyntax": false,
+		"lib": ["ESNext", "DOM"],
+		"preserveConstEnums": true,
+		"strict": true,
+		"outDir": "../../.moon/cache/types/libs/utilities"
+	},
+	"exclude": []
+}
