+++ tsconfig.json
@@ -16,6 +16,9 @@
     },
     {
       "path": "libs/graphql"
+    },
+    {
+      "path": "libs/utilities"
     }
   ]
 }
